LayoutTests/js/dom/domjit-function-effect-should-overlap-with-call-expected.txt

Test function calls should overlap with effects of DOMJIT functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test() is true
PASS successfullyParsed is true

TEST COMPLETE
